rename reusable prompt actions folder
