+++ src/vs/workbench/contrib/chat/browser/actions/reusablePromptActions/chatAttachPromptAction.ts
